Struct lde::ext::OpCodeBuilder

pub struct OpCodeBuilder {
    // some fields omitted
}

Unstable helper to construct opcodes from bytes.

 let _ = OpCode(OpCodeBuilder::new(5).write(0, 0x05u8).write(1, 42));

Methods

impl OpCodeBuilder

fn new(len: u8) -> OpCodeBuilder

Create a new opcode of specified length, max 15.

fn write<T: Copy>(&mut self, offset: usize, val: T) -> &mut OpCodeBuilder

Write a value to the opcode buffer at specified offset.