Trait lde::InsnSet

pub trait InsnSet: Clone {
    type Va: Copy + Clone + AddAssign + Add<Output=Self::Va>;
    fn ld(codes: &[u8]) -> u32;
    fn iter(codes: &[u8], va: Self::Va) -> LDIter<Self>;
}

Declares the entry point for an instruction set's length disassembler.

Associated Types

type Va: Copy + Clone + AddAssign + Add<Output=Self::Va>

Virtual address type.

Required Methods

fn ld(codes: &[u8]) -> u32

Length disassemble the given bytes.

Returns 0 on failure.

fn iter(codes: &[u8], va: Self::Va) -> LDIter<Self>

Create an iterator over the opcodes contained in the byte slice.

Provide a Virtual Address to keep track of the instruction pointer.